condense route modules
