+++ README.md
@@ -101,9 +101,8 @@ Additional environment variables:
 - `app.py` – application factory
 - `wsgi.py` – WSGI entry point for production servers
 - `service.py` – Windows service wrapper
-- `routes/` – blueprints and route handlers
-- The `diagrams` blueprint exposes `/diagrams` via the `diagrams.diagrams_index`
-  endpoint and `/diagram_converter` for the converter tool.
+- `routes/` – all blueprints live in `routes/core.py` with additional routes in `all_routes.py`.
+- The diagrams viewer and converter remain under the `diagrams` blueprint in `core.py`.
 - `templates/` – Jinja2 templates
 - `static/` – static assets
 - `config/` – default configuration files
